Refactored server class and iteration 0 is functional
